chore: move update stronghold files to outside of login folder
