+++ packages/desktop/views/index.js
@@ -1,5 +1,6 @@
 export * from './dashboard'
 export * from './login'
 export * from './onboarding'
+export * from './update-stronghold'
 
 export { default as Splash } from './Splash.svelte'
