+++ packages/desktop/views/login/views/index.js
@@ -1,5 +1,3 @@
-export * from './update-stronghold'
-
 export { default as EnterPinView } from './EnterPinView.svelte'
 export { default as LoadProfileView } from './LoadProfileView.svelte'
 export { default as SelectProfileView } from './SelectProfileView.svelte'
