+++ packages/desktop/views/update-stronghold/index.js
