+++ packages/desktop/views/update-stronghold/views/UpdateBackupView.svelte
