+++ packages/desktop/views/update-stronghold/views/UpdateStrongholdView.svelte
