+++ packages/desktop/views/update-stronghold/views/index.js
