+++ packages/shared/lib/core/router/enums/index.ts
@@ -1,4 +1,3 @@
-export * from './login'
 export * from './onboarding'
 export * from './settings'
 
@@ -8,3 +7,4 @@ export * from './dashboard-route.enum'
 export * from './governance-route.enum'
 export * from './login-route.enum'
 export * from './router-manager-extension-name.enum'
+export * from './update-stronghold-route.enum'
