--- packages/shared/lib/core/router/enums/login/index.ts
@@ -1,1 +1,0 @@
-export * from './update-stronghold-route.enum'
