+++ packages/shared/lib/core/router/enums/update-stronghold-route.enum.ts
