+++ packages/shared/lib/core/router/subrouters/index.ts
@@ -1,4 +1,4 @@
-export * from './login'
 export * from './onboarding'
 
 export * from './login-router'
+export * from './update-stronghold-router'
