--- packages/shared/lib/core/router/subrouters/login/index.ts
@@ -1,1 +1,0 @@
-export * from './update-stronghold-router'
